Add prometheus status
