+++ app/controllers/system_status.py
@@ -1,6 +1,6 @@
 
 from flask_restful import Resource
-from app.helpers.kube import get_cluster_status_info
+from app.controllers.statuses import get_cluster_status_info, get_prometheus_status_info
 from app.models.clusters import Cluster
 
 
@@ -9,7 +9,10 @@ class SystemStatusView(Resource):
     def get(self):
         # get clusters status
         clusters = Cluster.find_all()
-        clusters_status = {'clusters': get_cluster_status_info(clusters)}
-        print(clusters_status)
-
-        return dict(status='success', data=clusters_status), 200
+        clusters_status = get_cluster_status_info(clusters)
+        prometheus_status = get_prometheus_status_info(clusters)
+        
+        return dict(status='success', data={
+            'clusters_status': clusters_status,
+            'prometheus_status': prometheus_status
+        }), 200
